+++ src/app/app.jsx
@@ -1,0 +1,25 @@
+import logo from '../../assets/logo.svg';
+import './styles/app.scss';
+
+function App() {
+	return (
+		<div className="app">
+			<header className="app-header">
+				<img src={logo} className="app-logo" alt="logo" />
+				<p>
+					Edit <code>src/App.js</code> and save to reload.
+				</p>
+				<a
+					className="app-link"
+					href="https://reactjs.org"
+					target="_blank"
+					rel="noopener noreferrer"
+				>
+					Learn React
+				</a>
+			</header>
+		</div>
+	);
+}
+
+export default App;
