--- src/components/Order/Order.jsx
@@ -1,2 +1,0 @@
-// для блока Main
-// кнопка оформить заказ
