--- src/components/Сommercial/Commercial.jsx
@@ -1,1 +1,0 @@
-// блок рекламы GooglePlay и AppStore
