+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.scss';
-import App from './components/App/App';
+import App from './app/app';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
