+++ src/widget/deferred/deferred.jsx
