+++ src/widget/details/details.jsx
