+++ src/widget/navigation/navigation.jsx
