+++ src/widget/payment/payment.scss
