+++ src/widget/proof/proof.jsx
