+++ src/widget/rate/rate.jsx
