+++ src/widget/register/register.jsx
